Extract a TopAppBar.kt file inside DS
