+++ app/design-system/design-system-hedvig/src/main/kotlin/com/hedvig/android/design/system/hedvig/tokens/ScaffoldTokens.kt
@@ -1,10 +1,5 @@
 package com.hedvig.android.design.system.hedvig.tokens
 
-import androidx.compose.ui.unit.dp
-
 internal object ScaffoldTokens {
   val BackgroundColor = ColorSchemeKeyTokens.BackgroundPrimary
-  val TopAppBarHeight = 64.dp
-  val TopAppBarHorizontalPadding = 16.dp
-  val TopAppBarTextStyle = TypographyKeyTokens.HeadlineSmall
 }
