+++ app/design-system/design-system-hedvig/src/main/kotlin/com/hedvig/android/design/system/hedvig/tokens/TopAppBarTokens.kt
@@ -1,0 +1,11 @@
+package com.hedvig.android.design.system.hedvig.tokens
+
+import androidx.compose.ui.unit.dp
+
+internal object TopAppBarTokens {
+  val ContainerColor = ColorSchemeKeyTokens.BackgroundPrimary
+  val ContentColor = ColorSchemeKeyTokens.TextPrimary
+  val ContainerHeight = 64.dp
+  val ContentHorizontalPadding = 16.dp
+  val TextStyle = TypographyKeyTokens.HeadlineSmall
+}
