+++ micro-apps/design-showcase/src/main/kotlin/com/hedvig/android/sample/design/showcase/topbar/TopAppBar.kt
@@ -61,7 +61,8 @@ fun TopAppBarShowcase() {
       },
     ) {
       Column(
-        modifier = Modifier.fillMaxSize().border(1.dp, Color.Red)
+        modifier = Modifier
+          .fillMaxSize()
           .verticalScroll(rememberScrollState()),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally,
